move radio button list
