+++ src/packages/property-editors/radio-button-list/manifests.ts
