+++ src/packages/property-editors/radio-button-list/property-editor-ui-radio-button-list.element.ts
@@ -1,11 +1,11 @@
-import type { UmbInputRadioButtonListElement } from '../../../components/input-radio-button-list/input-radio-button-list.element.js';
+import type { UmbInputRadioButtonListElement } from '../../core/components/input-radio-button-list/input-radio-button-list.element.js';
 import { html, customElement, property, state } from '@umbraco-cms/backoffice/external/lit';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
 import { UmbPropertyValueChangeEvent } from '@umbraco-cms/backoffice/property-editor';
 import type { UmbPropertyEditorConfigCollection } from '@umbraco-cms/backoffice/property-editor';
 import type { UmbPropertyEditorUiElement } from '@umbraco-cms/backoffice/extension-registry';
 
-import '../../../components/input-radio-button-list/input-radio-button-list.element.js';
+import '../../core/components/input-radio-button-list/input-radio-button-list.element.js';
 
 /**
  * @element umb-property-editor-ui-radio-button-list
